pyproject.toml
[tool.poetry]
name = "twitter bot"
version = "0.1.0"
description = ""
authors = ["TheM1N9 <manikanta9k7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tweepy = "^4.14.0"
requests = "^2.32.3"
twikit = "^2.1.3"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
